update: ストリーミング機能を追加
