+++ nextjs-dashboard/app/dashboard/(overview)/loading.tsx
@@ -1,0 +1,5 @@
+import DashboardSkeleton from "../../ui/skeletons";
+
+export default function Loading() {
+    return <DashboardSkeleton />;
+  }
